billullo/pyproject.toml
[project]
name = "billullo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "django>=6.0.1",
    "djangorestframework>=3.16.1",
]
